toolchanger: Support for tools
Setting up toolchangers with Klipper is currently a free-for-all of macros and plugins. This is an attempt to introduce some common structure that is generic enough to fit almost all use cases while making most configs look familiar.

Design goals:

- Seamless integration with existing "always enabled" Klipper features. Like offsets, pressure advance, widely recognized GCodes.
- Enough flexibility to add optional features, like input shaper by the user.
- Tool changer hardware dev should be able to develop a base config that can be included as separate file without modifications.
- If two toolchange solutions are chainable physically, their configs should be chainable out of the box - provide structure to encourage self-contained configs.

https://github.com/TypQxQ/Klipper_ToolChanger is a capable implementation of something similar.
This is trying to be significantly more minimalistic and less opinionated.

Signed-off-by: Viesturs Zariņš <viesturz@gmail.com>